fix(vm-runner): fix Go test failures from workspace path sandboxing
FileServer.resolvePath() correctly rejects absolute paths and confines
access to the workspace root, but tests were passing absolute tmpdir
paths. Make FileServer root configurable via NewFileServerWithRoot() for
testing, update all test helpers to use relative paths, fix shell-escape
assertion mismatches in exec tests, and fix int64/int type mismatch in
VNC proxy test.

+++ openclaw-vm-runner/internal/server/vnc_proxy_test.go
@@ -189,5 +189,5 @@ func TestVNCProxy_BearerTokenAuth(t *testing.T) {
 
 func TestVNCProxy_ReadLimit(t *testing.T) {
 	// Verify the read limit constant is 1 MiB
-	assert.Equal(t, int64(1<<20), vncReadLimit)
+	assert.Equal(t, 1<<20, vncReadLimit)
 }
